+++ app/backend/lib/template-upload.ts
@@ -1,10 +1,14 @@
 import { Router } from 'express';
-import formidable from 'formidable';
+import formidable, { File } from 'formidable';
 import fs from 'fs';
 import * as XLSX from 'xlsx';
 import limiter from './excel_import/excel-limiter';
 import getAuthRole from '../../utils/getAuthRole';
-import { ExpressMiddleware, parseForm } from './express-helper';
+import {
+  ExpressMiddleware,
+  commonFormidableConfig,
+  parseForm,
+} from './express-helper';
 import readTemplateOneData from './excel_import/template_one';
 import readTemplateTwoData from './excel_import/template_two';
 
@@ -28,7 +32,7 @@ const processTemplateUpload: ExpressMiddleware = async (req, res) => {
   }
 
   const errorList = [];
-  const form = new formidable.IncomingForm({ maxFileSize: 8000000 });
+  const form = formidable(commonFormidableConfig);
 
   const files = await parseForm(form, req).catch((err) => {
     errorList.push({ level: 'file', error: err });
@@ -36,7 +40,9 @@ const processTemplateUpload: ExpressMiddleware = async (req, res) => {
   });
 
   const filename = Object.keys(files)[0];
-  const uploaded = files[filename];
+  const uploadedFilesArray = files[filename] as Array<File>;
+
+  const uploaded = uploadedFilesArray?.[0];
   if (!uploaded) {
     return res.status(200).end();
   }